Rename type to models for rest api
